requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
asyncpg==0.29.0
paho-mqtt==2.0.0
requests==2.31.0
